@@ -6,7 +6,7 @@ import android.content.Intent;
 import android.content.IntentFilter;
 import android.graphics.Color;
 import android.os.Parcelable;
-import android.support.v4.content.LocalBroadcastManager;
+import androidx.localbroadcastmanager.content.LocalBroadcastManager;
 import android.text.SpannableString;
 import android.text.Spanned;
 import android.text.style.ForegroundColorSpan;
@@ -24,6 +24,8 @@ import ohi.andre.consolelauncher.commands.main.raw.location;
 import ohi.andre.consolelauncher.commands.main.specific.RedirectCommand;
 import ohi.andre.consolelauncher.managers.AliasManager;
 import ohi.andre.consolelauncher.managers.AppsManager;
+import ohi.andre.consolelauncher.managers.LaunchInfo;
+import ohi.andre.consolelauncher.managers.AppUtils;
 import ohi.andre.consolelauncher.managers.ChangelogManager;
 import ohi.andre.consolelauncher.managers.ContactManager;
 import ohi.andre.consolelauncher.managers.HTMLExtractManager;
@@ -155,7 +157,8 @@ public class MainManager {
             Tuils.log(e);
         }
 
-        appsManager = new AppsManager(c);
+        appsManager = new AppsManager(mContext);
+
         aliasManager = new AliasManager(mContext);
 
         final OkHttpClient client = new OkHttpClient.Builder()
@@ -197,7 +200,7 @@ public class MainManager {
 //        }.start();
 
         rssManager = new RssManager(mContext, client);
-        themeManager = new ThemeManager(client, mContext, c);
+        themeManager = new ThemeManager(client, mContext, mContext);
         musicManager2 = XMLPrefsManager.getBoolean(Behavior.enable_music) ? new MusicManager2(mContext) : null;
         ChangelogManager.printLog(mContext, client);
         htmlExtractManager = new HTMLExtractManager(mContext, client);
@@ -242,14 +245,14 @@ public class MainManager {
                         LocalBroadcastManager.getInstance(context.getApplicationContext()).sendBroadcast(i);
                     }
 
-                    if(p != null && p instanceof AppsManager.LaunchInfo) {
-                        onCommand(cmd, (AppsManager.LaunchInfo) p, intent.getBooleanExtra(MainManager.MUSIC_SERVICE, false));
+                    if(p != null && p instanceof LaunchInfo) {
+                        onCommand(cmd, (LaunchInfo) p, intent.getBooleanExtra(MainManager.MUSIC_SERVICE, false));
                     } else {
                         onCommand(cmd, aliasName, intent.getBooleanExtra(MainManager.MUSIC_SERVICE, false));
                     }
                 } else if(action.equals(location.ACTION_LOCATION_CMD_GOT)) {
                     Tuils.sendOutput(context, "Lat: " + intent.getDoubleExtra(TuiLocationManager.LATITUDE, 0) + "; Long: " + intent.getDoubleExtra(TuiLocationManager.LONGITUDE, 0));
-                    TuiLocationManager.instance(context).rm(location.ACTION_LOCATION_CMD_GOT);
+                    TuiLocationManager.create(context).rm(location.ACTION_LOCATION_CMD_GOT);
                 }
             }
         };
@@ -272,7 +275,7 @@ public class MainManager {
         }
     }
 
-    public void onCommand(String input, AppsManager.LaunchInfo launchInfo, boolean wasMusicService) {
+    public void onCommand(String input, LaunchInfo launchInfo, boolean wasMusicService) {
         if(launchInfo == null) {
             onCommand(input, (String) null, wasMusicService);
             return;
@@ -394,7 +397,7 @@ public class MainManager {
 
     public CommandExecuter executer() {
         return (input, obj) -> {
-            AppsManager.LaunchInfo li = obj instanceof AppsManager.LaunchInfo ? (AppsManager.LaunchInfo) obj : null;
+            LaunchInfo li = obj instanceof LaunchInfo ? (LaunchInfo) obj : null;
 
             onCommand(input, li, false);
         };
@@ -408,7 +411,7 @@ public class MainManager {
     Pattern pp = Pattern.compile("%p", Pattern.CASE_INSENSITIVE | Pattern.LITERAL);
     Pattern pl = Pattern.compile("%l", Pattern.CASE_INSENSITIVE | Pattern.LITERAL);
 
-    public boolean performLaunch(MainPack mainPack, AppsManager.LaunchInfo i, String input) {
+    public boolean performLaunch(MainPack mainPack, LaunchInfo i, String input) {
         Intent intent = appsManager.getIntent(i);
         if (intent == null) {
             return false;
@@ -428,7 +431,7 @@ public class MainManager {
 
             SpannableString text = new SpannableString(a);
             text.setSpan(new ForegroundColorSpan(outputColor), 0, text.length(), Spanned.SPAN_EXCLUSIVE_EXCLUSIVE);
-            CharSequence s = TimeManager.instance.replace(text);
+            CharSequence s = TimeManager.instance(mContext).replace(text);
 
             Tuils.sendOutput(mainPack, s, TerminalManager.CATEGORY_OUTPUT);
         }
@@ -480,12 +483,12 @@ public class MainManager {
                 input = null;
             }
 
-            List<? extends Group> appGroups = info.appsManager.groups;
+            List<? extends Group> appGroups = (List<? extends Group>) (List<?>) info.appsManager.groups;
             if(appGroups != null) {
                 for(Group g : appGroups) {
                     if(name.equals(g.name())) {
                         if(input == null) {
-                            Tuils.sendOutput(mContext, AppsManager.AppUtils.printApps(AppsManager.AppUtils.labelList((List<AppsManager.LaunchInfo>) g.members(), false)));
+                            Tuils.sendOutput(mContext, AppUtils.printApps(AppUtils.labelList((List<LaunchInfo>) g.members(), false)));
                             return true;
                         } else {
                             return g.use(mainPack, input);
@@ -543,7 +546,7 @@ public class MainManager {
 
         @Override
         public boolean trigger(MainPack info, String input) {
-            AppsManager.LaunchInfo i = appsManager.findLaunchInfoWithLabel(input, AppsManager.SHOWN_APPS);
+            LaunchInfo i = appsManager.findLaunchInfoWithLabel(input, AppsManager.SHOWN_APPS);
             return i != null && performLaunch(info, i, input);
         }
     }
