@@ -13,8 +13,8 @@ import android.os.Binder;
 import android.os.Build;
 import android.os.IBinder;
 import android.os.PowerManager;
-import android.support.v4.app.NotificationCompat;
-import android.support.v4.app.RemoteInput;
+import androidx.core.app.NotificationCompat;
+import androidx.core.app.RemoteInput;
 
 import java.io.IOException;
 import java.util.Collections;
@@ -189,7 +189,7 @@ public class MusicService extends Service implements
         }
 
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.JELLY_BEAN) not = builder.build();
-        else not = builder.getNotification();
+        else not = builder.build();
 
         return not;
     }
