@@ -3,8 +3,9 @@ package ohi.andre.consolelauncher.tuils;
 import android.content.Context;
 import android.graphics.Canvas;
 import android.util.AttributeSet;
+import androidx.appcompat.widget.AppCompatEditText;
 
-public class OutlineEditText extends android.support.v7.widget.AppCompatEditText {
+public class OutlineEditText extends AppCompatEditText {
 
     private int drawTimes = -1;
 
